Updated forces to doubles, with the new values.  Haven't finished need to set for new force, and edited force.

+++ src/main/java/edu/odu/icat/analytics/AnalyticsAlgorithm.java
@@ -137,7 +137,7 @@ public abstract class AnalyticsAlgorithm implements  Runnable
             return controllabilityCheck.isSelected();
         }
 
-        public void addEntityToReport(Entity entity, int Degree)
+        public void addEntityToReport(Entity entity, double Degree)
         {
             ReportData.add(entity, Degree);
         }
